aho-corasick: use to_printable
